Fix possible divide by zero error
Also, add tests for this case.

Resolves: #98
