+++ app/src/test/java/org/fedorahosted/freeotp/TokenCodeTest.java
@@ -1,16 +1,12 @@
 package org.fedorahosted.freeotp;
 
-import android.net.Uri;
-
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mockito;
 import org.mockito.runners.MockitoJUnitRunner;
 
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.assertEquals;
-import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
 public class TokenCodeTest {
@@ -20,10 +16,10 @@ public class TokenCodeTest {
     @Before
     public void setUp() throws Exception {
         // otpauth://totp/FreeOTP:joe@google.com?secret=JBSWY3DPEHPK3PXP&issuer=FreeOTP
-        totpToken = mockToken("totp", null);
+        totpToken = TokenTestUtils.mockToken("totp", null);
 
         // otpauth://hotp/FreeOTP:joe@google.com?secret=JBSWY3DPEHPK3PXP&issuer=FreeOTP
-        hotpToken = mockToken("hotp", "0");
+        hotpToken = TokenTestUtils.mockToken("hotp", "0");
     }
 
     @Test
@@ -54,31 +50,4 @@ public class TokenCodeTest {
         assertEquals("282760", code);
     }
 
-    // Who tests the test utils?
-    public static Token mockToken(String authority, String counter){
-       return mockToken(authority, counter, "FreeOTP:joe@google.com");
-    }
-
-    public static Token mockToken(String authority, String counter, String id) {
-        // https://github.com/google/google-authenticator/wiki/Key-Uri-Format
-        Uri mMockUri = Mockito.mock(Uri.class);
-
-        when(mMockUri.getScheme()).thenReturn("otpauth");
-        when(mMockUri.getAuthority()).thenReturn(authority);
-        when(mMockUri.getPath()).thenReturn(id);
-        when(mMockUri.getQueryParameter("issuer")).thenReturn("FreeOTP");
-        when(mMockUri.getQueryParameter("secret")).thenReturn("JBSWY3DPEHPK3PXP");
-
-        if (authority.equals("hotp")) {
-            when(mMockUri.getQueryParameter("counter")).thenReturn(counter);
-        }
-
-        try {
-            return new Token(mMockUri);
-        } catch (Token.TokenUriInvalidException e) {
-            e.printStackTrace();
-        }
-
-        return null;
-    }
 }
